Disabling warning by casting in indel simulator.
git-svn-id: http://svn.seqan.de/seqan/trunk@8606 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/indel_simulator/indel_simulator.cpp
@@ -92,10 +92,10 @@ int main(int argc, const char *argv[])
 	while (!_streamEOF(file))
 	{
 		parse_skipWhitespace(file,c);
-		int rangeBegin = parse_readDouble(file,c);
+		int rangeBegin = static_cast<int>(parse_readDouble(file,c));
 		if(abs(rangeBegin) > absMaxValue) absMaxValue = abs(rangeBegin);
 		parse_skipWhitespace(file,c);
-		int rangeEnd = parse_readDouble(file,c);
+		int rangeEnd = static_cast<int>(parse_readDouble(file,c));
 		if(abs(rangeEnd) > absMaxValue) absMaxValue = abs(rangeEnd);
 		appendValue(options.ranges,Pair<int,int>(rangeBegin,rangeEnd));
 	}
